Add support for `@ParamConverter` in `LinkRequestListener`
The `@ParamConverter` annotation, and other request variable modifying
annotations are often used in REST APIs.
Most, if not all, of those annotations hook on to the
`kernel.controller` event to do their jobs.
By emitting such event in the `LinkRequestListener`, the arguments to
the controller are modified, and the developer does not have to care
if their actions are called by this listener, or from an external request.