Fix LinkRequestListener support for symfony 3, make tests work with symfony 3
Event::getDispatcher() is deprecated since symfony 2.4 and removed in symfony 3.0.
It is passed as 3rd parameter to the eventlistener starting at symfony 2.4.
Because the minimum required version for this bundle is symfony 2.1, Event::getDispatcher()
is still called when the listener did not receive a 3rd parameter.

Some settings in routing.yml have changed in symfony 3:
 * the pattern option is replaced by path.
 * requirements._method is replaced by methods.
These new options are not available in symfony 2.1, so the test configuration was
split in two trees, one for symfony 2 and one for symfony 3.

Newer versions of the sensio/framework-extra-bundle are required to be able
to use symfony 3, this changed the path of the ParamConverter class which is
hardcoded in Tests/bootstrap.php. Both locations are now checked.